redis 预减库存 + rabbitmq 异步下单

+++ src/main/java/com/csranger/miaosha/dao/OrderDao.java
@@ -25,12 +25,10 @@ public interface OrderDao {
      * statement 填入将会被执行的 SQL 字符串数组，keyProperty 填入将会被更新的参数对象的属性的值，before 填入 true 或 false 以
      * 指明 SQL 语句应被在插入语句的之前还是之后执行。resultType 填入 keyProperty 的 Java 类型
      */
-
-
     @Insert("insert into order_info(user_id, goods_id, delivery_addr_id, goods_name, goods_count, goods_price, order_channel, status, create_date) " +
             "values(#{userId}, #{goodsId}, #{deliveryAddrId}, #{goodsName}, #{goodsCount}, #{goodsPrice}, #{orderChannel}, #{status}, #{createDate})")
     @SelectKey(statement = "select last_insert_id()", keyProperty = "id", keyColumn = "id", resultType = long.class, before = false)
-    long insert(OrderInfo orderInfo);
+    int insert(OrderInfo orderInfo);
 
 
     /**
